feat: add metadata context and normalizeTwitterHandle utilities, update layout and metadata files
