+++ src/app/(spaces)/homebase/[[...slug]]/layout.tsx
@@ -40,7 +40,11 @@ function resolveCastEmbedImageUrl(embeds?: Embed[]): string | undefined {
   return undefined;
 }
 
-export async function generateMetadata({ params }): Promise<Metadata> {
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ slug?: string[] }>;
+}): Promise<Metadata> {
   const systemConfig = await loadSystemConfig();
   const baseUrl = await resolveBaseUrl({ systemConfig });
   const brandName = systemConfig.brand.displayName;
