+++ src/common/lib/utils/metadataContext.ts
@@ -1,0 +1,5 @@
+export type MetadataContext = {
+  baseUrl: string;
+  brandName: string;
+  twitterHandle?: string;
+};
