+++ src/common/lib/utils/tokenMetadata.tsx
@@ -1,12 +1,8 @@
 import { merge } from "lodash";
 import { Metadata } from "next";
 import { WEBSITE_URL } from "@/constants/app";
-
-type MetadataContext = {
-  baseUrl: string;
-  brandName: string;
-  twitterHandle?: string;
-};
+import type { MetadataContext } from "@/common/lib/utils/metadataContext";
+import { normalizeTwitterHandle } from "@/common/lib/utils/normalizeTwitterHandle";
 
 export type TokenMetadata = {
   name?: string;
@@ -89,22 +85,4 @@ export const getTokenMetadataStructure = (
   }
 
   return metadata;
-};
-
-const normalizeTwitterHandle = (handle?: string): string | undefined => {
-  if (!handle) {
-    return undefined;
-  }
-  const trimmed = handle.trim();
-  if (!trimmed) {
-    return undefined;
-  }
-  const withoutAt = trimmed.startsWith("@") ? trimmed.slice(1) : trimmed;
-  const cleaned = withoutAt.replace(/^https?:\/\//i, "");
-  const parts = cleaned.split("/");
-  const lastPart = parts[parts.length - 1]?.replace(/^@/, "");
-  if (!lastPart) {
-    return undefined;
-  }
-  return `@${lastPart}`;
 };
